+++ src/Operation/State/SortLimited/BufferNotFull.php
@@ -1,0 +1,51 @@
+<?php declare(strict_types=1);
+
+namespace FiiSoft\Jackdaw\Operation\State\SortLimited;
+
+use FiiSoft\Jackdaw\Internal\Item;
+use FiiSoft\Jackdaw\Operation\SortLimited;
+use SplHeap;
+
+final class BufferNotFull extends State
+{
+    private int $length;
+    private int $count = 0;
+    
+    public function __construct(SortLimited $operation, SplHeap $buffer, int $length)
+    {
+        parent::__construct($operation, $buffer);
+        
+        $this->length = $length;
+    }
+    
+    public function hold(Item $item): void
+    {
+        $this->buffer->insert($item->copy());
+    
+        if (++$this->count === $this->length) {
+            $this->bufferFull();
+        }
+    }
+    
+    public function setLength(int $length): void
+    {
+        if ($length !== $this->length) {
+            if ($length < $this->count) {
+                do {
+                    $this->buffer->extract();
+                } while (--$this->count > $length);
+            }
+            
+            $this->length = $length;
+    
+            if ($this->count === $this->length) {
+                $this->bufferFull();
+            }
+        }
+    }
+    
+    private function bufferFull(): void
+    {
+        $this->operation->transitTo(new BufferFull($this->operation, $this->buffer));
+    }
+}