+++ src/Operation/Strategy/Unique/Strategy.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FiiSoft\Jackdaw\Operation\Strategy\Unique;
+
+use FiiSoft\Jackdaw\Internal\Item;
+
+interface Strategy
+{
+    public function check(Item $item): bool;
+}